notifier/requirements.txt
yoyo-migrations==8.1.0
PyMySQL==1.0.2